pyproject.toml
[project]
name = "chat-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
